feat(pci-instances): move api types to dedicated folder
